+++ app/Http/Controllers/RentController.php
@@ -29,7 +29,7 @@ class RentController extends Controller
                 return view('newRentByType')->with(['type' => $type, 'vehicles' => $vehicles, 'erreur']);
             }
         } else {
-            return view('newRent');
+            return redirect()->action('RentController@getPage');
         }
     }
 
@@ -39,11 +39,18 @@ class RentController extends Controller
             $loc = Auth::user()->vehicle;
             $vehicle_id = $id;
             if ($loc == 0) {
-                $vehicle = Vehicle::where('id', $id)->get();
-                return view('newRentByCar')->with(['vehicle' => $vehicle, 'id' => $vehicle_id]);
+                $vehicle = Vehicle::where('id', $id)->where('disponibilite', 1)->get();
+                if (count($vehicle) == 0) {
+                    return redirect()->action('RentController@getPage');
+                }
+                else{
+
+                return view('newRentByCar')->with(['vehicle' => $vehicle, 'id' => $vehicle_id]);                  
+                }
+
             } else {
                 return back();
-            }
+            } 
         } else {
             return redirect()->action('ToolsController@connexion');
         }
@@ -51,16 +58,23 @@ class RentController extends Controller
 
     public function saveRent(Request $request)
     {
+        $request->validate([
+            'dateDepart'=> 'required|date|after_or_equal: today',
+            'dateArrive'=> 'required|date|after_or_equal:dateDepart'
+        ]);
+
         $rent = [
             'user_id' => Auth::user()->id,
             'vehicle_id' => $request->input('id'),
-            'dateDepart' => $request->input('datedebut'),
-            'dateArrive' => $request->input('datefin')
+            'dateDepart' => $request->input('dateDepart'),
+            'dateArrive' => $request->input('dateArrive')
         ];
         Rent::create($rent);
 
         $user = User::where('id', Auth::user()->id);
         $user->update(['vehicle' => 1]);
+        $alert = "Votre réservation à été pris en compte. Toutefois, les dates de réservation sont succeptible d'\être modifier";
+        $request->session()->flash('alert_mp', $alert);
         return redirect()->action('ToolsController@main');
     }
 
@@ -72,13 +86,25 @@ class RentController extends Controller
             return view('allRent')->with(['myRents' => $myRents]);
         }
         else {
-            return view('allRent');
+            return redirect()->action('ToolsController@connexion');
         }
     }
 
     public function getRent(){
-        $rents = Rent::where('created_at', '!=', null)->with('vehicle', 'user')->get();
-        return view('rent')->with(['rents'=> $rents]);
+
+        if (Auth::check()) {
+            if (Auth::user()->admin) {
+                
+                $rents = Rent::where('created_at', '!=', null)->with('vehicle', 'user')->get();
+                return view('rent')->with(['rents'=> $rents]);
+            }
+            else {
+                return redirect()->action('ToolsController@main');   
+            }
+        }
+        else {
+            return redirect()->action('ToolsController@main');   
+        }
     }
 
     public function getCurrentRent()
@@ -89,7 +115,7 @@ class RentController extends Controller
             return view('currentRent')->with(['myRents' => $myRents, 'date' => $date]);
         }
         else {
-            return view('currentRent');
+            return redirect()->action('ToolsController@connexion');
         }
         
     }
@@ -97,22 +123,34 @@ class RentController extends Controller
 
 
     public function delete($id, Request $request){
-        $alert = 'Votre suppression sur une location à été réalisée avec succès';
-        $request->session()->flash('alert_mp', $alert);
-        $rent = Rent::where('id', $id)->with('user')->first();
-        
-        $user = [
-           'nom'=> $rent->user->nom,
-           'prenom'=> $rent->user->prenom,
-           'mail'=> $rent->user->mail,
-           'vehicle'=> 0,
-           'admin'=> $rent->user->admin
-        ];
-        User::where('id', $rent->vehicle_id)->first()->update($user);
 
-        Rent::where('id', $id)->delete();
-
+        if (Auth::check()) {
+            if (Auth::user()->admin) {
+                
+                $alert = 'Votre suppression sur une location à été réalisée avec succès';
+                $request->session()->flash('alert_mp', $alert);
+                $rent = Rent::where('id', $id)->with('user')->first();
+                
+                $user = [
+                   'nom'=> $rent->user->nom,
+                   'prenom'=> $rent->user->prenom,
+                   'mail'=> $rent->user->mail,
+                   'vehicle'=> 0,
+                   'admin'=> $rent->user->admin
+                ];
+                User::where('id', $rent->vehicle_id)->first()->update($user);
         
-        return redirect()->action('ToolsController@main');
+                Rent::where('id', $id)->delete();
+        
+                
+                return redirect()->action('ToolsController@main');
+            }
+            else {
+                return redirect()->action('ToolsController@main');   
+            }
+        }
+        else {
+            return redirect()->action('ToolsController@main');   
+        }
     }
 }
